requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-cors-headers==4.1.0
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
GDAL==3.6.2
MarkupPy==1.14
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
numpy @ file:///C:/b/abs_5akk51tu0f/croot/numpy_and_numpy_base_1687466253743/work
odfpy==1.4.1
openpyxl==3.1.2
psycopg2-binary==2.9.6
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
